Move HyperParameters module into deprecated file. Add deprecation warnings
